Move implementations of TypeChainElement methods to new cpp file
